+++ examples/shared_samples/logical_enclosure.rb
@@ -1,4 +1,4 @@
-# (C) Copyright 2020 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2021 Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # You may not use this file except in compliance with the License.
@@ -20,11 +20,11 @@ require_relative '../_client' # Gives access to @client
 # - C7000 and Synergy for all API versions
 
 # Resource Class used in this sample
-logical_enclosure_class = OneviewSDK.resource_named('LogicalEnclosure', @client.api_version)
+logical_enclosure_class = OneviewSDK.resource_named('LogicalEnclosure', @client.api_version, 'Synergy')
 
 # Scope class used in this sample
-encl_group_class = OneviewSDK.resource_named('EnclosureGroup', @client.api_version)
-enclosure_class = OneviewSDK.resource_named('Enclosure', @client.api_version)
+encl_group_class = OneviewSDK.resource_named('EnclosureGroup', @client.api_version, 'Synergy')
+enclosure_class = OneviewSDK.resource_named('Enclosure', @client.api_version, 'Synergy')
 
 variant = 'Synergy'