+++ examples/shared_samples/os_deployment_plan.rb
@@ -1,4 +1,4 @@
-# (C) Copyright 2017 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2021 Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # You may not use this file except in compliance with the License.
@@ -11,8 +11,8 @@
 
 require_relative '../_client' # Gives access to @client
 
-# Supported APIs:
-# - 300, 500, 600, 800, 1000, 1200, 1600, 1800, 2000 only for Synergy
+# Supported Variants
+# Synergy for all api versions
 
 # Resource Class used in this sample
 os_deployment_plan_class = OneviewSDK.resource_named('OSDeploymentPlan', @client.api_version, 'Synergy')