+++ spec/shared_context.rb
@@ -129,6 +129,15 @@ RSpec.shared_context 'integration api2200 context', a: :b do
   end
 end
 
+# Context for API2400 integration testing:
+RSpec.shared_context 'integration api2400 context', a: :b do
+  before :all do
+    integration_context
+    $client_2400 ||= OneviewSDK::Client.new($config.merge(api_version: 2400))
+    $client_2400_synergy ||= OneviewSDK::Client.new($config_synergy.merge(api_version: 2400))
+  end
+end
+
 # Context for Image Streamer API300 integration testing:
 RSpec.shared_context 'integration i3s api300 context', a: :b do
   before :all do
@@ -233,6 +242,13 @@ RSpec.shared_context 'system api2200 context', a: :b do
   before(:each) do
     load_system_properties
     generate_clients(2200)
+  end
+end
+
+RSpec.shared_context 'system api2400 context', a: :b do
+  before(:each) do
+    load_system_properties
+    generate_clients(2400)
   end
 end
 
@@ -367,6 +383,9 @@ def generate_clients(api_version)
   when 2200
     $client_2200 ||= OneviewSDK::Client.new($config.merge(api_version: api_version))
     $client_2200_synergy ||= OneviewSDK::Client.new($config_synergy.merge(api_version: api_version))
+  when 2400
+    $client_2400 ||= OneviewSDK::Client.new($config.merge(api_version: api_version))
+    $client_2400_synergy ||= OneviewSDK::Client.new($config_synergy.merge(api_version: api_version))
   end
 
   allow_any_instance_of(OneviewSDK::Client).to receive(:appliance_api_version).and_call_original